Makefile outside working directory, o files live outside LLAMAS directory, recovery test done, injectFault.h deleted
