+++ Arduino/LLAMAS/LLAMAS.ino
@@ -1,6 +1,5 @@
 
 #include "recovery.h"
-// #include "pid.h"
 
 void setup() {
 }
